Add a min-height to the label in .field-inputs
Checkboxes and radio buttons are put in the .field-inputs container as
with other form elements but since they are floated, they do not hold
open the container themselves. Rather, the text in the surrounding
label is used to hold open div.field-inputs and this is critical to
maintaining consistent spacing between the form elements.

 -------------   ------------------------
| Field label | | [] Checkbox label text |
 -------------   ------------------------
  ^-floated       ^- floated   ^- text holds open non-floated label

 -------------   ----
| Field label | | [] |
 -------------   ----
  ^- all the elements here are floated -
     nothing is holding the checkbox label
     open and this row collapses into the
     one beneath

By adding a min height to the label, it will act as if it has text
even if it is empty. The min-height has been chosen to match the
line-height value of 15px, as specified on the body.